move SQLModels models to infra layer
this was never meant to be our actual domain model, but temporarially.
