+++ src/__setup__.py
@@ -68,7 +68,7 @@ def setup_database(vitae: VitaeSettings) -> None:
     """Setups database."""
     # ``models`` module must be evaluated before create or drop it.
     # That is why this imports an unused variable inside this function.
-    from src import models  # noqa: F401, PLC0415
+    from src.infra.database import models  # noqa: F401
 
     if vitae.in_development:
         # Since the dataset for development is far smaller than the production
